mcptools/requirements.txt
mcp>=1.26.0
python-dotenv>=1.0.0
PyJWT[crypto]>=2.8.0
httpx>=0.27.0
uvicorn>=0.29.0
